% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindCoders.R
\name{bindCoders}
\alias{bindCoders}
\title{Creates a table with events from the image level annotation information,
 with an additional columns with the names of the coders.}
\usage{
bindCoders(wearableCamImagesList, namesList = NULL, minDuration = 1)
}
\arguments{
\item{wearableCamImagesList}{a list of \code{wearableCamImages} objects.}

\item{namesList}{(optional) a list of names for the coders. It must be the same length as wearableCamImagesList
and contains different names.
 @param minDuration the minimal number of images for defining an event. Default is 1.}
}
\value{
A \code{tbl_df} with event index, start time (POSIXt),
end time (POSIXt), annotation (character) and coder name.
}
\description{
Creates a table with events from the image level annotation information,
 with an additional columns with the names of the coders.
}
\examples{
data("dummyWearableCamImages")
bindCoders(list(dummyWearableCamImages, dummyWearableCamImages), minDuration = 1)
}

