% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-methods.R
\docType{methods}
\name{codes}
\alias{codes}
\alias{codes,codes-method}
\alias{codes,wearableCamImages-method}
\title{Accesses codes}
\usage{
codes(wearableCamImages)

\S4method{codes}{wearableCamImages}(wearableCamImages)
}
\arguments{
\item{wearableCamImages}{the wearableCamImages object}
}
\description{
Accesses codes
}

