% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-methods.R
\docType{methods}
\name{codesBinaryVariables}
\alias{codesBinaryVariables}
\alias{codesBinaryVariables,codesBinaryVariables-method}
\alias{codesBinaryVariables,wearableCamImages-method}
\title{Accesses codesBinaryVariables}
\usage{
codesBinaryVariables(wearableCamImages)

\S4method{codesBinaryVariables}{wearableCamImages}(wearableCamImages)
}
\arguments{
\item{wearableCamImages}{the wearableCamImages object}
}
\description{
Accesses codesBinaryVariables
}

