% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{coding1}
\alias{coding1}
\title{A dummy WearableCamImages object for one coder.}
\format{A dummy WearableCamImages object.}
\description{
A dummy WearableCamImages object for one coder.
}
\examples{
data(coding1)
}
\keyword{data}

