% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{coding_example}
\alias{coding_example}
\title{A dummy WearableCamImages object.}
\format{A dummy WearableCamImages object.}
\description{
A dummy WearableCamImages object.
}
\examples{
data(coding_example)
}
\keyword{data}

