% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineObjects.R
\name{combineObjects}
\alias{combineObjects}
\title{Allows to merge several \code{wearableCamImages} objects,
i.e. for looking at coding results of one rater.}
\usage{
combineObjects(wearableCamImagesList)
}
\arguments{
\item{wearableCamImagesList}{a list of 2 \code{wearableCamImages} objects.}
}
\value{
A list of \code{wearableCamImages} objects.
}
\description{
Allows to merge several \code{wearableCamImages} objects,
i.e. for looking at coding results of one rater.
}
\examples{
data('dummyWearableCamImages')
wearableCamImagesList <- list(dummyWearableCamImages, dummyWearableCamImages)
correctedList <- combineObjects(wearableCamImagesList)
}

