% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCsvInput.R
\name{convertInput}
\alias{convertInput}
\title{Creates a wearableCamImages object from information read in a csv file.}
\usage{
convertInput(pathResults, sepResults, quoteSign = "'", pathDicoCoding,
  sepDicoCoding, formatDate = "ymd")
}
\arguments{
\item{pathResults}{the path to the file with coding results}

\item{sepResults}{the separator in the file with coding results}

\item{quoteSign}{the quote argument of read.table for the results, default is "\'"}

\item{pathDicoCoding}{the path to the file with the list of annotations}

\item{sepDicoCoding}{the separator in the file with the list of annotations}

\item{formatDate}{either 'ymd' or 'dmy'.}
}
\value{
A \code{wearableCamImages} object.
}
\description{
Creates a wearableCamImages object from information read in a csv file.
}
\details{
Please check the format that both files should have by looking at the provided
example data.
However you could consider write your own function for converting the input
instead of having to re-format all your existing data,
which we could even add to the package.
}
\examples{
pathResults <- system.file('extdata', 'image_level_pinocchio.csv', package = 'watchme')
sepResults <- ','
pathDicoCoding <-  system.file('extdata', 'dicoCoding_pinocchio.csv', package = 'watchme')
sepDicoCoding <- ';'
wearableCamImagesObject <- convertInput(pathResults=pathResults, sepResults=sepResults,
              pathDicoCoding=pathDicoCoding, sepDicoCoding=sepDicoCoding)
class(wearableCamImagesObject)
}

