% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-methods.R
\docType{methods}
\name{dicoCoding}
\alias{dicoCoding}
\alias{dicoCoding,dicoCoding-method}
\alias{dicoCoding,wearableCamImages-method}
\title{Accesses dicoCoding}
\usage{
dicoCoding(wearableCamImages)

\S4method{dicoCoding}{wearableCamImages}(wearableCamImages)
}
\arguments{
\item{wearableCamImages}{the wearableCamImages object}
}
\description{
Accesses dicoCoding
}

