% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dummyWearableCamImages}
\alias{dummyWearableCamImages}
\title{A dummy WearableCamImages object.}
\format{A dummy WearableCamImages object.}
\description{
A dummy WearableCamImages object.
}
\keyword{data}

