% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-methods.R
\docType{methods}
\name{imagePath}
\alias{imagePath}
\alias{imagePath,imagePath-method}
\alias{imagePath,wearableCamImages-method}
\title{Accesses imagePath}
\usage{
imagePath(wearableCamImages)

\S4method{imagePath}{wearableCamImages}(wearableCamImages)
}
\arguments{
\item{wearableCamImages}{the wearableCamImages object}
}
\description{
Accesses imagePath
}

