% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{IO2}
\alias{IO2}
\title{A dummy WearableCamImages object for one coder.}
\format{A dummy WearableCamImages object.}
\description{
A dummy WearableCamImages object for one coder.
}
\examples{
data(IO2)
}
\keyword{data}

