% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interraterReliability.R
\name{irrWatchme}
\alias{irrWatchme}
\title{Calculates interrater reliability using the irr package.}
\usage{
irrWatchme(wearableCamImagesList, namesList = NULL, oneToOne = FALSE,
  byGroup = FALSE, byCode = FALSE)
}
\arguments{
\item{wearableCamImagesList}{a list of \code{wearableCamImages} objects.}

\item{namesList}{(optional) a vector of names for the coders. It must be the same length as wearableCamImagesList
and contains different names.}

\item{oneToOne}{a boolean indicating whether Cohen's kappa should be calculated for each possible
pair of coders in case of more than 2 coders,
instead of Fleiss’s Kappa for all coders at the same time.}

\item{byGroup}{boolean indicating whether the IRR should be calculated for each group of codes separately.}

\item{byCode}{boolean indicating whether the IRR should be calculated for each code separately. If both
byGroup and byCode are FALSE annotations are compared as they are, so make sure all raters use e.g. the same separator
between annotations when they give several codes to the same picture.}
}
\value{
A \code{tbl_df} presenting the results of a call to the \code{irr} function.
If there are only two raters the called function is \code{kappa2}, unweighted.
 If there are more than two raters and \code{oneToOne} is \code{FALSE}, the called function is \code{kappam.fleiss}.
}
\description{
Calculates interrater reliability using the irr package.
}
\examples{
data("dummyWearableCamImages")
listWC <- list(dummyWearableCamImages, dummyWearableCamImages)
namesList <- c("Cain", "Abel")
irrWatchme(listWC, namesList=namesList)
listWC2 <- list(dummyWearableCamImages, dummyWearableCamImages, dummyWearableCamImages)
namesList <- c("Riri", "Fifi", "Loulou")
irrWatchme(listWC2, namesList=namesList)
irrWatchme(listWC2, namesList=namesList, oneToOne=TRUE)
irrWatchme(listWC, namesList=c("Cain", "Abel"), oneToOne=TRUE, byCode=TRUE)
irrWatchme(listWC, namesList=c("Cain", "Abel"), oneToOne=TRUE, byGroup=TRUE)
}

