% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{listWM}
\alias{listWM}
\title{A list of dummy WearableCamImages objects}
\format{A list of dummy WearableCamImages object.}
\description{
A list of dummy WearableCamImages objects,
several passes but same set of pictures
}
\examples{
data(listWM)
}
\keyword{data}

