% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputDifferences.R
\name{outputDifferences}
\alias{outputDifferences}
\title{Outputs images for which codes are different among coders}
\usage{
outputDifferences(wearableCamImagesList, namesList = NULL)
}
\arguments{
\item{wearableCamImagesList}{a list of \code{wearableCamImages} objects.}

\item{namesList}{(optional) a vector of names for the coders. It must be the same length as wearableCamImagesList
and contains different names.}
}
\value{
A \code{tbl_df} with image name, image time, and one column for each coder.
}
\description{
Outputs images for which codes are different among coders
}
\examples{
data('dummyWearableCamImages')
listWC <- list(dummyWearableCamImages, dummyWearableCamImages)
namesList <- c('Cain', 'Abel')
outputDifferences(listWC, namesList=namesList)
}

