% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-methods.R
\docType{methods}
\name{participantID}
\alias{participantID}
\alias{participantID,participantID-method}
\alias{participantID,wearableCamImages-method}
\title{Accesses participantID}
\usage{
participantID(wearableCamImages)

\S4method{participantID}{wearableCamImages}(wearableCamImages)
}
\arguments{
\item{wearableCamImages}{the wearableCamImages object}
}
\description{
Accesses participantID
}

