% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequence.R
\name{plotSequence}
\alias{plotSequence}
\title{Plot sequences of events}
\usage{
plotSequence(event_table, doNotUseCode = NULL, xAxis = "time",
  facettingGroup = FALSE, facettingCoder = FALSE, dicoCoding = NULL,
  cbbPaletteYes = TRUE)
}
\arguments{
\item{event_table}{a table of events created with the\code{toEventLevel} function (or having the same structure).}

\item{doNotUseCode}{a vector of codes that you do not want to see on the graph, e.g. if you have both codes
for categories and subcategories you may want to not plot categories.}

\item{xAxis}{either 'time' or 'picture' index as x axis variable.}

\item{facettingGroup}{boolean indicating whether there should be one plot per group of activities
(as defined in the dicoCoding)}

\item{facettingCoder}{boolean indicating whether there should be one plot per coder
(use only if you have a coder column!)}

\item{dicoCoding}{(optional) if you want consistent color definitons across several event tables, please provide the dicoCoding.}

\item{cbbPaletteYes}{boolean if dicoCoding provided and >=7 different activities, you can opt for a colorblind-friendly palette.}
}
\value{
A \code{ggplot2} graph.
}
\description{
Plot sequences of events
}
\examples{
data('dummyWearableCamImages')
library('ggplot2')
event_table <- toEventLevel(wearableCamImagesObject=dummyWearableCamImages)
plotSequence(event_table, dicoCoding = dummyWearableCamImages$dicoCoding)
plotSequence(event_table, xAxis='picture', facettingGroup=TRUE,
dicoCoding = dummyWearableCamImages$dicoCoding)
data('IO1')
data('IO2')
event_tableCoders <- bindCoders(list(IO1, IO2), minDuration = 1)
plotSequence(event_tableCoders, facettingGroup = TRUE, facettingCoder = TRUE,
dicoCoding=IO1$dicoCoding)
}

