% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryEventTable.R
\name{summaryEventTable}
\alias{summaryEventTable}
\title{Creates a table with summary statistics for event of each (sub)category.}
\usage{
summaryEventTable(eventTable)
}
\arguments{
\item{eventTable}{a table of events created with the\code{toEventLevel} function (or having the same structure).}
}
\value{
A data table with the (sub)category, the number of identified events, the average number of pictures, and the average duration in seconds.
}
\description{
Creates a table with summary statistics for event of each (sub)category.
}
\examples{
data("dummyWearableCamImages")
library("ggplot2")
eventTable <- toEventLevel(wearableCamImagesObject=dummyWearableCamImages)
summaryEventTable(eventTable)
}

