% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-methods.R
\docType{methods}
\name{timeDate}
\alias{timeDate}
\alias{timeDate,timeDate-method}
\alias{timeDate,wearableCamImages-method}
\title{Accesses timeDate}
\usage{
timeDate(wearableCamImages)

\S4method{timeDate}{wearableCamImages}(wearableCamImages)
}
\arguments{
\item{wearableCamImages}{the wearableCamImages object}
}
\description{
Accesses timeDate
}

