% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEventLevel.R
\name{toEventLevel}
\alias{toEventLevel}
\title{Creates a table with events from the image level annotation information.}
\usage{
toEventLevel(wearableCamImagesObject, minDuration = 1)
}
\arguments{
\item{wearableCamImagesObject}{the werableCamImagesObject contining the image level annotation information
(and the dico coding, of course)}

\item{minDuration}{the minimal number of images for defining an event. Default is 1.}
}
\value{
A \code{tbl_df} with event index, start time (POSIXt), end time (POSIXt) and eventCode (character).
}
\description{
Creates a table with events from the image level annotation information.
}
\examples{
data('dummyWearableCamImages')
eventTable <- toEventLevel(wearableCamImagesObject=dummyWearableCamImages)
eventTable
eventTable2 <- toEventLevel(wearableCamImagesObject=dummyWearableCamImages, minDuration = 2)
eventTable2
}

