% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEventLevel.R
\name{watchme_aggregate}
\alias{watchme_aggregate}
\title{Creates a table with events from the image level annotation information.}
\usage{
watchme_aggregate(df, min_no_pictures = 1)
}
\arguments{
\item{min_no_pictures}{the minimal number of images for defining an event. Default is 1.}

\item{wearableCamImagesObject}{the werableCamImagesObject contining the image level annotation information
(and the dico coding, of course)}
}
\value{
A \code{tbl_df} with event index, start time (POSIXt), end time (POSIXt) and event_code (character).
}
\description{
Creates a table with events from the image level annotation information.
}
\examples{
data('coding_example')
eventTable <- watchme_aggregate(df = coding_example)
eventTable
eventTable2 <- watchme_aggregate(df = coding_example, min_no_pictures = 2)
eventTable2
}

