% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineObjects.R
\name{watchme_combine_results}
\alias{watchme_combine_results}
\title{Combine objects with same pictures but different codes}
\usage{
watchme_combine_results(results_list, common_codes = c("non codable"))
}
\arguments{
\item{results_list}{a list of \code{tibble}  created by \code{watchme_prepare_data}.}

\item{common_codes}{One or several codes that are common between the objects.}
}
\value{
A \code{tibble} with
\itemize{
\item participantID Name or ID number of the participant (character)
\item image_time Path or name of the image in order to be able to identify duplicates (character)
\item image_time Time and date of each image (POSIXt)
\item booleanCodes columns of boolean, indicating if a given code was given to a given picture. codes is a condensed form of this.
\item the attribute \code{dico} \code{tibble} for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)
}
}
\description{
Combine objects with same pictures but different codes
}
\details{
The motivation was to create a single object based on several coding files
in the CHAI project: each pass/group of codes such as cooking and travel resulted
 in one file,
and we want to combine them.

For\code{common_codes} codes, the annotation are merged: if an image has such a code in an object
and not in the other, the code is assigned to the image in the resulting object.
The code does not look for conflicts.
}
\examples{
path_results <- system.file('extdata', 'oneday_CK.csv', package = 'watchme')
sep_results <- ','
path_dico <-  system.file('extdata', 'dico_pinocchio.csv', package = 'watchme')
sep_dico <- ';'
data_pictures <- watchme_prepare_data(path_results=path_results, sep_results=sep_results,
              path_dico=path_dico, sep_dico=sep_dico)
}

