% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interraterReliability.R
\name{watchme_ira}
\alias{watchme_ira}
\title{Calculates interrater agreement using the irr package. The unit of comparison is one picture.}
\usage{
watchme_ira(results_list, names_list = NULL, one_to_one = FALSE,
  byGroup = FALSE, by_code = FALSE)
}
\arguments{
\item{results_list}{a list of \code{tibble}  created by \code{watchme_prepare_data}.}

\item{names_list}{(optional) a vector of names for the coders. It must be the same length as results_list
and contains different names.}

\item{one_to_one}{a boolean indicating whether Cohen's kappa should be calculated for each possible
pair of coders in case of more than 2 coders,
instead of Fleiss's Kappa for all coders at the same time.}

\item{byGroup}{boolean indicating whether the IRR should be calculated for each group of codes separately. The meaning is, agreement = giving a code of the same group.}

\item{by_code}{boolean indicating whether the IRR should be calculated for each code separately. If both
byGroup and by_code are FALSE annotations are compared as they are.}
}
\value{
A \code{tbl_df} presenting the results of a call to the \code{irr} function.
If there are only two raters the called function is \code{kappa2}, unweighted.
 If there are more than two raters and \code{one_to_one} is \code{FALSE}, the called function is \code{kappam.fleiss}.
}
\description{
Calculates interrater agreement using the irr package. The unit of comparison is one picture.
}
\examples{
data('IO1')
data('IO2')
results_list <- list(IO1, IO2)
names_list <- c('Cain', 'Abel')
watchme_ira(results_list, names_list=names_list)
results_list2 <- list(IO1, IO1, IO2)
names_list <- c('Riri', 'Fifi', 'Loulou')
watchme_ira(results_list2, names_list=names_list)
watchme_ira(results_list2, names_list=names_list, one_to_one=TRUE)
watchme_ira(results_list, names_list=c('Cain', 'Abel'), one_to_one=TRUE, by_code=TRUE)
watchme_ira(results_list, names_list=c('Cain', 'Abel'), one_to_one=TRUE, byGroup=TRUE)
}

