% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputDifferences.R
\name{watchme_output_differences}
\alias{watchme_output_differences}
\title{Outputs images for which codes are different among coders}
\usage{
watchme_output_differences(results_list, names_list = NULL)
}
\arguments{
\item{results_list}{a list of \code{tibble} created by \code{watchme_prepare_data}.}

\item{names_list}{(optional) a vector of names for the coders. It must be the same length as results_list
and contains different names.}
}
\value{
A \code{tbl_df} with image name, image time, and one column for each coder.
}
\description{
Outputs images for which codes are different among coders
}
\examples{
\dontrun{
data('coding1')
data('coding2')
# With two coders
results_list <- list(coding1, coding2)
names_list <- c('Cain', 'Abel')
watchme_output_differences(results_list = results_list,
 names_list = names_list)
}
}
