% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_events.R
\name{watchme_plot_raw}
\alias{watchme_plot_raw}
\title{Plot}
\usage{
watchme_plot_raw(df)
}
\arguments{
\item{df}{}
}
\description{
Plot
}
\examples{
pathResults <- system.file('extdata', 'image_level_pinocchio.csv', package = 'watchme')
sepResults <- ','
pathDicoCoding <-  system.file('extdata', 'dicoCoding_pinocchio.csv', package = 'watchme')
sepDicoCoding <- ';'
data_pictures <- watchme_prepare_data(pathResults=pathResults, sepResults=sepResults,
              pathDicoCoding=pathDicoCoding, sepDicoCoding=sepDicoCoding)
watchme_plot_raw(data_pictures)
}

