% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequence.R
\name{watchme_plot_sequence}
\alias{watchme_plot_sequence}
\title{Plot sequences of events}
\usage{
watchme_plot_sequence(event_table, x_axis = "time")
}
\arguments{
\item{event_table}{a table of events created with the\code{toEventLevel} function (or having the same structure),
including dico as attribute.}

\item{x_axis}{either 'time' or 'picture' index as x axis variable.}
}
\value{
A \code{ggplot2} graph.
}
\description{
Plot sequences of events
}
\examples{
data('coding1')
library('ggplot2')
event_table <- watchme_aggregate(df = coding1)
watchme_plot_sequence(event_table)
watchme_plot_sequence(event_table, x_axis = "picture")
}
