% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCsvInput.R
\name{watchme_prepare_data}
\alias{watchme_prepare_data}
\title{Creates a wearableCamImages object from information read in a csv file.}
\usage{
watchme_prepare_data(path_results, sep_results, quote_sign = "'",
  participant_id = "no_id", path_dico, sep_dico, tz = "Asia/Kolkata")
}
\arguments{
\item{path_results}{the path to the file with coding results}

\item{sep_results}{the separator in the file with coding results}

\item{quote_sign}{the quote argument of read.table for the results, default is "\'"}

\item{participant_id}{participant ID, if available.}

\item{path_dico}{the path to the file with the list of annotations}

\item{sep_dico}{the separator in the file with the list of annotations}

\item{tz}{timezone}
}
\value{
A \code{tibble} with
\itemize{
\item participant_id Name or ID number of the participant (character)
\item image_time Path or name of the image in order to be able to identify duplicates (character)
\item image_time Time and date of each image (POSIXt)
\item booleanCodes columns of boolean, indicating if a given code was given to a given picture. codes is a condensed form of this.
\item the attribute \code{dico} \code{tibble} for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)
}
}
\description{
Creates a wearableCamImages object from information read in a csv file.
}
\details{
Please check the format that both files should have by looking at the provided
example data.
However you could consider write your own function for converting the input
instead of having to re-format all your existing data,
which we could even add to the package.
}
\examples{
path_results <- system.file('extdata', 'image_level_pinocchio.csv', package = 'watchme')
sep_results <- ','
path_dico <-  system.file('extdata', 'dicoCoding_pinocchio.csv', package = 'watchme')
sep_dico <- ';'
data_pictures <- watchme_prepare_data(path_results=path_results, sep_results=sep_results,
              path_dico=path_dico, sep_dico=sep_dico)
data_pictures
attr(data_pictures, "dico")
}

