% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-class.R
\docType{class}
\name{wearableCamImages-class}
\alias{wearableCamImages}
\alias{wearableCamImages-class}
\title{An S4 class to store information about images and their code.}
\description{
An S4 class to store information about images and their code.
}
\section{Slots}{

\describe{
\item{\code{participantID}}{Name or ID number of the participant (character)}

\item{\code{imagePath}}{Path or name of the image in order to be able to identify duplicates (character)}

\item{\code{timeDate}}{Time and date of each image (POSIXt)}

\item{\code{codes}}{annotation(s) given to this image (character), e.g. separated by ';'.}

\item{\code{codesBinaryVariables}}{table of boolean, indicating if a given code was given to a given picture. \code{codes} is a condensed
form of this slot.}

\item{\code{dicoCoding}}{table for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)}
}}

