% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wearableCamImages-class.R
\docType{class}
\name{wearableCamImages}
\alias{wearableCamImages}
\title{An R6 class to store information about images and their code.}
\format{\code{\link{R6Class}} object.}
\usage{
wearableCamImages
}
\value{
Object of \code{\link{R6Class}}.
}
\description{
An R6 class to store information about images and their code.
}
\section{Fields}{

\describe{
\item{\code{\code{participantID}}}{Name or ID number of the participant (character)}

\item{\code{\code{imagePath}}}{Path or name of the image in order to be able to identify duplicates (character)}

\item{\code{\code{timeDate}}}{Time and date of each image (POSIXt)}

\item{\code{\code{codes}}}{annotation(s) given to this image (character), e.g. separated by ','.}

\item{\code{\code{booleanCodes}}}{table of boolean, indicating if a given code was given to a given picture. codes is a condensed form of this slot.}

\item{\code{\code{dicoCoding}}}{table for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)}
}}
\examples{
pathResults <- system.file("extdata", "image_level_pinocchio.csv", package = "watchme")
sepResults <- ","
pathDicoCoding <-  system.file("extdata", "dicoCoding_pinocchio.csv", package = "watchme")
sepDicoCoding <- ";"
wearableCamImagesObject <- convertInput(pathResults=pathResults, sepResults=sepResults,
                                        pathDicoCoding=pathDicoCoding, sepDicoCoding=sepDicoCoding)
class(wearableCamImagesObject)
}
\keyword{data}

