% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineObjects.R
\name{combineObjects}
\alias{combineObjects}
\title{Combine objects with same pictures but different codes}
\usage{
combineObjects(wearableCamImagesList, codeException = c("non codable"))
}
\arguments{
\item{wearableCamImagesList}{a list of \code{wearableCamImages} objects.}

\item{codeException}{One or several codes that are common between the objects.}
}
\value{
A \code{wearableCamImages} object.
}
\description{
Combine objects with same pictures but different codes
}
\details{
The motivation was to create a single object based on several coding files
in the CHAI project: each pass/group of codes such as cooking and travel resulted in one file,
and we want to combine them.
For codeException codes, the annotation are merged: if an image has such a code in an object
and not in the other, the code is assigned to the image in the resulting object.
The code does not look for conflicts.
}
\examples{
data(listWM)
object <- combineObjects(listWM, codeException = c("non codable"))
}

