% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequence.R
\name{plotSequence}
\alias{plotSequence}
\title{Creates a table with events from the image level annotation information.}
\usage{
plotSequence(eventTable, doNotUseCode = NULL, xAxis = "time",
  facettingGroup = FALSE, facettingCoder = FALSE, dicoCoding = NULL,
  cbbPaletteYes = TRUE)
}
\arguments{
\item{eventTable}{a table of events created with the\code{toEventLevel} function (or having the same structure).}

\item{doNotUseCode}{a vector of codes that you do not want to see on the graph, e.g. if you have both codes
for categories and subcategories you may want to not plot categories.}

\item{xAxis}{either "time" or "picture" index as x axis variable.}

\item{facettingGroup}{boolean indicating whether there should be one plot per group of activities
(as defined in the dicoCoding)}

\item{facettingCoder}{boolean indicating whether there should be one plot per coder
(use only if you have a coder column!)}

\item{dicoCoding}{(optional) if you want consistent color definitons across several event tables, please provide the dicoCoding.}

\item{cbbPaletteYes}{boolean if dicoCoding provided and >=7 different activities, you can opt for a colorblind-friendly palette.}
}
\value{
A \code{ggplot2} graph.
}
\description{
Creates a table with events from the image level annotation information.
}
\examples{
data("dummyWearableCamImages")
library("ggplot2")
eventTable <- toEventLevel(wearableCamImagesObject=dummyWearableCamImages)
plotSequence(eventTable)
plotSequence(eventTable, xAxis="picture", facettingGroup=TRUE)
eventTableCoders <- bindCoders(list(dummyWearableCamImages, dummyWearableCamImages), minDuration = 1)
plotSequence(eventTableCoders, facettingGroup = TRUE, facettingCoder = TRUE,
dicoCoding=dummyWearableCamImages@dicoCoding)
}

