% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toEventLevel.R
\name{watchme_aggregate}
\alias{watchme_aggregate}
\title{Creates a table with events from the image level annotation information.}
\usage{
watchme_aggregate(df, min_no_pictures = 1)
}
\arguments{
\item{df}{the data created by using \code{watchme_prepare_data}}

\item{min_no_pictures}{the minimal number of images for defining an event. Default is 1.}
}
\value{
A \code{tbl_df} with
\itemize{
\item event code
\item start time (POSIXt)
\item end time (POSIXt)
\item no of pictures in the event
\item index of the first picture in the event
\item index of the last picture in the event
\item group of the code
\item meaning of the code
\item the attribute \code{dico} \code{tibble} for defining the codes with at least Code and Meaning column, possibly Group column for having groups of codes (e.g. sport encompasses running and swimming)
} event index, , ,  and event_code (character).
}
\description{
Creates a table with events from the image level annotation information.
}
\examples{
data('coding_example')
event_table <- watchme_aggregate(df = coding_example)
event_table
event_table2 <- watchme_aggregate(df = coding_example, min_no_pictures = 2)
event_table2
}

