% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_events.R
\name{watchme_plot_raw}
\alias{watchme_plot_raw}
\title{Plot}
\usage{
watchme_plot_raw(df)
}
\arguments{
\item{df}{A \code{tibble} created by \code{watchme_prepare_data}.}
}
\value{
a \code{ggplot}.
}
\description{
Plot
}
\examples{
data("coding_example")
watchme_plot_raw(coding_example)
}
